test: address cancellation readiness review feedback
